move mocks in dedicated package
